% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtempirical.R
\name{irt_empirical_model}
\alias{irt_empirical_model}
\title{}
\usage{
irt_empirical_model(data = NULL, items = NULL, addlogit = TRUE)
}
\arguments{
\item{data}{}

\item{items}{}
}
\description{

}
\examples{
sim_dichotomous<-data.frame(psych::sim.irt(nvar = 10, n = 1000, low=-3, high=3,a=NULL,c=0,z=1,d=NULL,mu=0,sd=1,mod="logistic")$items)
sim_polytomous<-data.frame(psych::sim.poly(nvar = 10 ,n = 1000, low=-2,high=2,a=NULL,c=0,z=1,d=NULL, mu=0,sd=1,cat=5,mod="logistic")$items)
irt_empirical_model(data=sim_dichotomous,items="V1")
irt_empirical_model(data=sim_polytomous,items="V1")
}
\keyword{assumptions}
